+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - The actor's prompt was incorrectly indicating an internal maximum of only one change per moment instead of one change per _line_.
+- The director's moment definition prompt was incorrectly asserting there was a previous moment even if there was not one.
+- Optimized `_execute_full_moment()`
 
 ## [0.1.2] - 2025-08-12